feat(routes): /ai-weekender filtered showcase route
Adds a new /ai-weekender route that lists only projects tagged
buildathon=ai-weekender, reusing the existing project list UI.
Extracts the inline list rendering from _index into a shared
ProjectListView component so both the main feed and the new
showcase route share one source of truth (sort tabs, voting,
infinite scroll, host pick, skeleton, empty state).

- New ProjectListView component (components/ProjectListView.tsx)
  takes headerTitle/headerSubtitle, optional buildathonFilter,
  optional emptyState, and a refreshKey prop.
- _index.tsx now mounts ProjectListView for its main feed and
  bumps refreshKey after a successful submit. Submit modal,
  ActivityFeed sidebar, and page chrome are unchanged.
- ai-weekender.tsx renders ProjectListView with
  buildathonFilter='ai-weekender', a small page title, a one-line
  subtitle, and route-level meta tags for SEO. Vanilla list per
  decision locked in handoff.
